+++ lib/app.dart
@@ -5,8 +5,10 @@ import 'package:app/pages/account/create/backupAccountPage.dart';
 import 'package:app/pages/account/create/createAccountPage.dart';
 import 'package:app/pages/account/createAccountEntryPage.dart';
 import 'package:app/pages/account/import/importAccountPage.dart';
+import 'package:app/pages/assets/announcementPage.dart';
 import 'package:app/pages/assets/asset/assetPage.dart';
 import 'package:app/pages/assets/asset/locksDetailPage.dart';
+import 'package:app/pages/assets/manage/manageAssetsPage.dart';
 import 'package:app/pages/assets/transfer/detailPage.dart';
 import 'package:app/pages/assets/transfer/transferPage.dart';
 import 'package:app/pages/homePage.dart';
@@ -52,6 +54,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:get_storage/get_storage.dart';
+// import 'package:jpush_flutter/jpush_flutter.dart';
 import 'package:polkawallet_sdk/api/types/networkParams.dart';
 import 'package:polkawallet_sdk/api/types/walletConnect/pairingData.dart';
 import 'package:polkawallet_sdk/api/types/walletConnect/payloadData.dart';
@@ -250,11 +253,10 @@ class _WalletAppState extends State<WalletApp> {
     _startPlugin(service);
   }
 
-  Future<void> _changeToKusamaForKar() async {
-    final name = 'kusama';
+  Future<void> _switchNetwork(String networkName) async {
     await _changeNetwork(
-        widget.plugins.firstWhere((e) => e.basic.name == name));
-    _service.store.assets.loadCache(_keyring.current, name);
+        widget.plugins.firstWhere((e) => e.basic.name == networkName));
+    _service.store.assets.loadCache(_keyring.current, networkName);
   }
 
   Future<void> _changeNode(NetworkParams node) async {
@@ -374,6 +376,10 @@ class _WalletAppState extends State<WalletApp> {
             : null,
       );
 
+      if (_keyring.keyPairs.length > 0) {
+        _store.assets.loadCache(_keyring.current, _service.plugin.basic.name);
+      }
+
       _startPlugin(service);
     }
 
@@ -400,7 +406,7 @@ class _WalletAppState extends State<WalletApp> {
                     if (snapshot.hasData && _service != null) {
                       return snapshot.data > 0
                           ? HomePage(_service, _connectedNode,
-                              _checkJSCodeUpdate, _changeToKusamaForKar)
+                              _checkJSCodeUpdate, _switchNetwork)
                           : CreateAccountEntryPage();
                     } else {
                       return Container(color: Theme.of(context).canvasColor);
@@ -445,6 +451,8 @@ class _WalletAppState extends State<WalletApp> {
       TransferDetailPage.route: (_) => TransferDetailPage(_service),
       TransferPage.route: (_) => TransferPage(_service),
       LocksDetailPage.route: (_) => LocksDetailPage(_service),
+      ManageAssetsPage.route: (_) => ManageAssetsPage(_service),
+      AnnouncementPage.route: (_) => AnnouncementPage(),
 
       /// profile
       SignMessagePage.route: (_) => SignMessagePage(_service),